refactor(semantic): flatten `src/semantic/assembly` module structure
In accordance with Razor's essentialism philosophy (KISS, YAGNI, DRY), this commit eliminates the unnecessary directory nesting within the `semantic` module. The `src/semantic/assembly/mod.rs` and `src/semantic/assembly/model.rs` files, which were tightly coupled via DTOs (`AssembledStatement`, `Constituent`, etc.) have been merged into a single `src/semantic/assembly.rs` file. This reduces architectural layers and cognitive overhead without altering compiler behavior.
